aligning DTOs and repo with backend upgrade

+++ android/app/src/main/java/com/r2r/readytorevise/data/remote/dto/ReviewDto.kt
@@ -7,7 +7,7 @@ data class ReviewDto(
     val id: String,
     val studyItemId: String,
     val result: String,
-    val intervalDays: Int,
+    val intervalDays: Double,
     val reviewedAt: String,
     val nextReviewAt: String,
     val isCorrect: Boolean? = null,
